update purchases ui

+++ src/packages/purchases/modules/fuel/components/NewFuelPurchase.jsx
@@ -19,6 +19,7 @@ import { apiFetchUtil, GetGross } from "@/utils";
 import WebStorage from "@/utils/WebStorage";
 import { APPNAME } from "@/environments";
 import DataGridToolbar from "../../../shared/components/DataGridToolbar";
+import { MdOutlineSaveAlt } from "react-icons/md";
 
 
 const NewFuelPurchase = () => {
@@ -186,15 +187,32 @@ const NewFuelPurchase = () => {
         field: 'actions',
         type: 'actions',
         width: 80,
-        getActions: (params) => [
-          <GridActionsCellItem
-            key={uuidv4()}
-            icon={<MdDelete size={25} />}
-            label="Delete"
-            onClick={deleteItem(params.id)}
-          />,
-        ],
         hideable: false,
+        getActions: (params) => {
+          const isInEditMode = rowModesModel[params.id]?.mode === GridRowModes.Edit;
+
+          if (isInEditMode) {
+            return [
+              <GridActionsCellItem 
+              key={uuidv4()}
+              icon={<MdOutlineSaveAlt />}
+              label="Save"
+              sx={{
+                color: 'primary.main',
+              }}
+              onClick={handleSaveClick(params.id)}
+              />
+            ]
+          }
+          [
+            <GridActionsCellItem
+              key={uuidv4()}
+              icon={<MdDelete size={25} />}
+              label="Delete"
+              onClick={deleteItem(params.id)}
+            />,
+          ]
+        },
       },
     ],
     [deleteItem],
